feat: Vote Page UI
