+++ src/@components/VotePage/VoteContent/style.ts
@@ -12,9 +12,32 @@ const VoteContentTitle = styled.h2`
   color: ${({ theme }) => theme.colors.gray900};
 `;
 
+const VoteBtnContainer = styled.div`
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+
+  display: flex;
+  justify-content: center;
+  gap: 0.1rem;
+`;
+
+const VoteBtn = styled.button`
+  width: 100%;
+  height: 5.8rem;
+
+  ${({ theme }) => theme.fonts.btn3}
+
+  background-color: ${({ theme }) => theme.colors.gray900};
+  color: ${({ theme }) => theme.colors.white};
+`;
+
 const St = {
   VoteContentContainer,
   VoteContentTitle,
+  VoteBtnContainer,
+  VoteBtn,
 };
 
 export default St;
